MAT-6075: Refactor cql graph creation (running the translator) into abstract class
